Support tree leafs and non-leafs at same level
Previously the behavior of having a leaf and non-leaf object at the
same level of the tree was undefined.  With this commit we make this
defined behavior.  We update the tree.get() function to return the
leaf node, any non-leaf nodes, and the found node object.

Currently we're using a special key, "__leaf", to designate a leaf
node, but this is subject to change.  We might make this a gensym.